requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-docx>=1.1.0
pypdf>=3.17.0
